Fix infinite loop problem.
Problem: Program went into an infinite loop
when inputting more than 1 word into occupation section of data entry
this is due to the standard library “cin” function not flushing stdin
and failing to convert a non-integer ASCII value to an integer
which sets the error flag which is not cleared automatically.

Solution: To fix “cin” from its error condition,
check for error after the conversion to integer,
and clear the error if so error is present.
To prevent extra data in stdin, flush stdin
after getting a good conversion from “cin”.